+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 ## Unreleased
 
 
+### Added
+
+-   use [nodejs-support-dates](https://github.com/jokeyrhyme/nodejs-support-dates.js) to provide warnings about unsupported versions
+
+
 ## 1.0.1 - 2016-10-25